[ty] Refine ty doc prototype rendering
